# swagger.yaml
openapi: 3.0.0
info:
  title: Dynamic API
  version: 1.0.0
...
paths:
...
  /newsletter:
    post:
      tags:
        - newsletter
      summary: Insert new newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/newsletter'
      responses:
        '200':
          description: newsletter inserted successfully
    get:
      tags:
        - newsletter
      summary: Get all newsletters (filtered)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: Filter query
      responses:
        '200':
          description: List of newsletters
    put:
      tags:
        - newsletter
      summary: Update existing newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/newsletter'
      responses:
        '200':
          description: newsletter updated successfully
    delete:
      tags:
        - newsletter
      summary: Delete newsletter by filter
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: Filter criteria for delete
      responses:
        '200':
          description: newsletter deleted successfully
  /newsletter/{id}:
    get:
      tags:
        - newsletter
      summary: Get newsletter by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: newsletter found
        '404':
          description: newsletter not found
...
components:
  schemas:
...
    newsletter:
      type: object
      properties:
        newsletterId:
          type: string
        email:
          type: string
...